add integration test & refactor
